Adding handling for numerous email recipients
In the sendEmailForm component, if too many recipients are provided then
it would overflow the pill container.  Additionally, there was no way to
see any additional details about the recipients when they were shown as
pills.

With this change, the number of pills shown is capped to a maximum and
if there are more recipients than that a single pill is shown with a
count of the number of recipients.

A new button is added to the send email form which will open a modal
dialog listing all recipients, with additional information (like their
email address).  The information available in this dialog may increase
in the future (i.e. show the recipient type, opt out information, etc).
Another future change may be to introduce the ability to modify the
recipient list via this dialog.