Send email updates
Updating sending of SingleEmailMessages to support the use of merge
field syntax in the body of the message.  While there is not currently a
UI for inserting this merge syntax into the message while composing, at
the very least we can support using the merge fields if the user enters
them through some other mechanism (or manually) while composing the
message in the UI.  This will be more useful later when/if there is a UI
to insert merge syntax via the UI.

Adding logic to show a success message and close the modal dialog upon
successful sending of the email (which currently is interpted to mean
there wasn't an exception thrown by the Apex controller, which will need
to be updated in the future to allow more sophisticated email send
success/failure reporting).