Adding 'emails' directory to sfdx-project.json
Since the default package directory was changed in sfdx-project.json to
only refer to the `force-app/volunteers` directory, this was excluding
changes made to metadata in the `force-app/emails` directory from being
included when deploying the project using sfdx.

This change adds the `force-app/emails` directory as an additional
package directory so that it is included when deploying the project
using sfdx, while the `force-app/volunteers` directory is maintained as
the default package directory.