Adding check for Contact.Email FLS to email functionality
Previously, in the custom send email action, it was assumed that if no
recipient returned in the recipient list had an address property or if
all addresses were blank, then it was because the user did not have FLS
access to the Contact.Email field.

This assumption would then be used to determine the behavior of the UI
for the user, such as not including informative messages which relate to
the email address of the recipients (i.e. notifying the user when some of
the recipients do not have an email address before sending the email, or
including the 'Email' column in displays of recipients).

However, this logic was flawed in that it falsly assumed that the user
did not have FLS access to the Contact.Email field if they happened to
be emailing a list of recipients where *none* of the recipients had a
valid email address.

To resolve this issue, now an explicit check for FLS access to the
Contact.Email field is performed, rather than trying to infer access to
this field based on the data returned.  The result of the explicit FLS
check is used to control whether the user is given any information about
email addresses, including warning messages and whether the Email column
appears in displays.

Note that in either case, the user never was able to access the Email
address of the Contact if they didn't have FLS access, because the data
was never (and is still not) returned to the client.  What was failing
and is now fixed is that read access to that field is now more reliably
determined which makes the UI more consistent.