Updating email delivery report
Updating email delivery report to include the number of emails which
were successfully sent.

Since this information wasn't actually available, EmailController was
updated to include explicitly the number of successful emails which were
sent in the EmailResults return value.  Additionally, an explicit boolean
was added to EmailResults to indicate whether errors exist, which makes
it easier to check whether any errors exist rather than awkwardly trying
to count the number of keys in the errors map which is returned.

Now, instead of taking as a parameter the list of errors, the email
delivery report accepts the email result object (which contains the list
of errors and now the number of successes).

Logic which detects whether there were any errors when sending emails or
show the number of emails successfully sent have been updated to use the
two new properties returned as part of the EmailResults.