Adding check for user email verification
Users are not allowed to send emails unless the email address they are
using to send the email is "verified" (which usually means the the user
has clicked a link sent to that address to verify the email address
belongs to them).

We don't want to allow users to attempt to send emails using a
non-verified address (which won't work), so this change updates the
"From" address options in the send emails components to only allow the
user to select from "verified" email addresses.  We warn the user if
there are no "verified" email addresses available that they may need to
verify their email address.

Note that this does mean the user does not need to have a verified email
address themselves in the case where there are verified org-wide email
addresses available to choose from.

Additionally, this change introduces support for the Sender Name and
Sender Email options (which more closely matches how the native send
email action works).  This is a feature which allows a user to override
the name and email address used when sending emails to be different from
the normal name and email values on their user record.  It is important
to support these features both because it is more accurate to display
the actual values used when sending the email, but also because there
are different rules for verified email addresses when using Sender
Email-- specifically that a user's own email address does not need to be
verified if they are using a Sender Email address, but that Sender Email
address must be verified before it is allowed to be used.

Many code changes were made to support the above changes, including the
fact that the EmailAddress class now indicates the verification status
of the email address represented (where this is relevant), the list of
"from" address options sent to the sendEmailsAction component now
filtering on this verified status, and changes to sendEmailsAction and
sendEmailsForm components to remove assumptions around the "from"
options always containing at least one valid value (which was the user's
email address) and now these components are set up to better handle
variations in the "from" options list returned from Apex.